Server/requirements.txt
pywin32
pycryptodome
construct
servicemanager
validators
netifaces